Add initial implementation of an http api.
This attempt at an http api uses python-falcon as a WSGI and Routing
framework. There is still a lot of work that needs to be done but it works
with some of the basics listed in the api spec. Namely, api revisions, CA
listing, and certificate requests. User auth is just handled by a simple
implementation using htpasswd files and passlib.

A new 'serve' command is added to the janus-cli, which is meant for local
access to the CAs. This command will start a server using python-eventlet and
handle requests. Eventlet is not listed as an install requirement since the
command function detects its absence and reports appropriately. This should
probably be the way falcon and passlib should be handled or eventlet changed
to an install dependency. This is yet to be determined.

Lots of docs are still needed. Testing has only been done with curl one-offs.